Final NChooseR...
Should be the final version of program. Due to Java's limitations in storing large int values, I had to create Band-Aid solutions to allow for the program to work correctly. For example, the value of "n" cannot ever be smaller than "r", as it leads to value being calculated to be too large for it to be stored as a long data value (>9,223,372,036,854,775,807), causing incorrect results, but NO CRASHING (phewwww...). Also, the largest value that can be given in the program is 20! DO NOT MAKE IT LARGER, as Java CANNOT store that in a long data format and will cause it to create a "divide by 0" error (memory wipe).